Permission Issue - Edit / Delete - Done

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -50,6 +50,28 @@ public class ImageController {
         return "images/image";
     }
 
+    //This method is called when a non owner of an image tries to edit an image.
+    @RequestMapping("/images/{id}/error")
+    public String showImageIdError(@PathVariable("id") Integer id, Model model) {
+        Image image = imageService.getImageById(id);
+        model.addAttribute("image", image);
+        model.addAttribute("tags", image.getTags());
+        String error = "Only the owner of the image can edit the image";
+        model.addAttribute("editError",error);
+        return "images/image";
+    }
+
+    //This method is called when a non owner of an image tries to delete an image.
+    @RequestMapping("/images/{id}/deleteerror")
+    public String showImageIdDeleteError(@PathVariable("id") Integer id, Model model) {
+        Image image = imageService.getImageById(id);
+        model.addAttribute("image", image);
+        model.addAttribute("tags", image.getTags());
+        String error = "Only the owner of the image can delete the image";
+        model.addAttribute("deleteError",error);
+        return "images/image";
+    }
+
     //This controller method is called when the request pattern is of type 'images/upload'
     //The method returns 'images/upload.html' file
     @RequestMapping("/images/upload")
@@ -89,13 +111,18 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model,HttpSession session) {
         Image image = imageService.getImage(imageId);
-
-        String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        User user = (User) session.getAttribute("loggeduser");
+        if(user.getId() != image.getUser().getId() ){
+            return "redirect:images/"+image.getId()+"/error";
+        }
+        else {
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -129,7 +156,7 @@ public class ImageController {
         updatedImage.setDate(new Date());
 
         imageService.updateImage(updatedImage);
-        return "redirect:/images/" + updatedImage.getTitle();
+        return "redirect:/images/" + updatedImage.getId();
     }
 
 
@@ -137,9 +164,15 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,HttpSession session) {
+
+        User user = (User)session.getAttribute("loggeduser");
+        if(user.getId() != imageService.getImage(imageId).getUser().getId() ){
+            return "redirect:images/"+ imageService.getImage(imageId).getId()+"/deleteerror";
+        }else{
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
     }
